requirements.txt
jupyterlab
datasets
torch
transformers
pillow
ipywidgets
torchvision
git+https://github.com/openai/CLIP.git
diffusers
accelerate
comet_ml
loguru